Add files via upload
This file contains a number of functions written in R which convert a large list of patient results which is presented as a table where each row is a single result, into a data table where each row represents one time for one patient and each column represents one type of result.

It was written for an AI project based in two UK quaternary hospitals. It is unfinished; complete version coming soon.